Move CqlValue and Row to scylla_cql::value module
Previously those types were located in scylla_cql::frame::response::result.
This is because in old framework they were strongly coupled to the
result frame: values were always deserialized into `Vec` of `Row`s, which
in turn were `Vec`s of `CqlValue`.

In the new framework this coupling is gone. New traits allow deserialization
from raw bytes instead of `CqlValue`, which means that those types are
no more important than any other, and can be moved to the module with
other values.

scylla_cql::values may be a bit of a weird place for a `Row` struct,
but I don't see any better place.
