+++ examples/basic.rs
@@ -3,7 +3,7 @@ use futures::StreamExt as _;
 use futures::TryStreamExt as _;
 use scylla::client::session::Session;
 use scylla::client::session_builder::SessionBuilder;
-use scylla::frame::response::result::Row;
+use scylla::value::Row;
 use scylla::DeserializeRow;
 use std::env;
