+++ examples/cql-time-types.rs
@@ -6,7 +6,7 @@ use chrono::{DateTime, NaiveDate, NaiveTime, Utc};
 use futures::{StreamExt as _, TryStreamExt as _};
 use scylla::client::session::Session;
 use scylla::client::session_builder::SessionBuilder;
-use scylla::frame::response::result::CqlValue;
+use scylla::value::CqlValue;
 use scylla::value::{CqlDate, CqlTime, CqlTimestamp};
 use std::env;
