+++ examples/tower.rs
@@ -1,6 +1,6 @@
 use scylla::client::session::Session;
 use scylla::client::session_builder::SessionBuilder;
-use scylla::frame::response::result::Row;
+use scylla::value::Row;
 use std::env;
 use std::future::Future;
 use std::pin::Pin;
@@ -43,11 +43,9 @@ async fn main() -> anyhow::Result<()> {
         .await?
         .into_rows_result()?;
 
-    let print_text = |t: &Option<scylla::frame::response::result::CqlValue>| {
+    let print_text = |t: &Option<scylla::value::CqlValue>| {
         t.as_ref()
-            .unwrap_or(&scylla::frame::response::result::CqlValue::Text(
-                "<null>".to_string(),
-            ))
+            .unwrap_or(&scylla::value::CqlValue::Text("<null>".to_string()))
             .as_text()
             .unwrap_or(&"<null>".to_string())
             .clone()
