+++ scylla-cql/src/deserialize/value_tests.rs
@@ -13,14 +13,12 @@ use std::sync::Arc;
 use crate::deserialize::value::{TupleDeserializationErrorKind, TupleTypeCheckErrorKind};
 use crate::deserialize::{DeserializationError, FrameSlice, TypeCheckError};
 use crate::frame::response::result::NativeType::*;
-use crate::frame::response::result::{
-    CollectionType, ColumnType, CqlValue, NativeType, UserDefinedType,
-};
+use crate::frame::response::result::{CollectionType, ColumnType, NativeType, UserDefinedType};
 use crate::serialize::value::SerializeValue;
 use crate::serialize::CellWriter;
 use crate::value::{
     Counter, CqlDate, CqlDecimal, CqlDecimalBorrowed, CqlDuration, CqlTime, CqlTimestamp,
-    CqlTimeuuid, CqlVarint, CqlVarintBorrowed,
+    CqlTimeuuid, CqlValue, CqlVarint, CqlVarintBorrowed,
 };
 
 use super::{
