+++ scylla-cql/src/serialize/value_tests.rs
@@ -1,6 +1,4 @@
-use crate::frame::response::result::{
-    CollectionType, ColumnType, CqlValue, NativeType, UserDefinedType,
-};
+use crate::frame::response::result::{CollectionType, ColumnType, NativeType, UserDefinedType};
 use crate::macros::SerializeValue;
 use crate::serialize::value::{
     BuiltinSerializationError, BuiltinSerializationErrorKind, BuiltinTypeCheckError,
@@ -10,7 +8,8 @@ use crate::serialize::value::{
 };
 use crate::serialize::{CellWriter, SerializationError};
 use crate::value::{
-    Counter, CqlDate, CqlDuration, CqlTime, CqlTimestamp, CqlTimeuuid, CqlVarint, MaybeUnset, Unset,
+    Counter, CqlDate, CqlDuration, CqlTime, CqlTimestamp, CqlTimeuuid, CqlValue, CqlVarint,
+    MaybeUnset, Unset,
 };
 
 use std::collections::hash_map::DefaultHasher;
