+++ scylla/src/lib.rs
@@ -125,8 +125,8 @@ pub mod frame {
 
             pub(crate) use scylla_cql::frame::response::result::*;
             pub use scylla_cql::frame::response::result::{
-                CollectionType, ColumnSpec, ColumnType, CqlValue, NativeType, PartitionKeyIndex,
-                Row, TableSpec, UserDefinedType,
+                CollectionType, ColumnSpec, ColumnType, NativeType, PartitionKeyIndex, TableSpec,
+                UserDefinedType,
             };
         }
     }
