+++ scylla/src/observability/driver_tracing.rs
@@ -3,9 +3,10 @@ use crate::network::Connection;
 use crate::response::query_result::QueryResult;
 use crate::routing::{Shard, Token};
 use crate::utils::pretty::{CommaSeparatedDisplayer, CqlValueDisplayer};
+use crate::value::deser_cql_value;
 use itertools::Either;
+use scylla_cql::frame::response::result::ColumnSpec;
 use scylla_cql::frame::response::result::RawMetadataAndRawRows;
-use scylla_cql::frame::response::result::{deser_cql_value, ColumnSpec};
 use std::borrow::Borrow;
 use std::fmt::Display;
 use std::sync::atomic::AtomicUsize;
