+++ scylla/tests/integration/cql_value.rs
@@ -1,8 +1,7 @@
 use assert_matches::assert_matches;
 
 use scylla::client::session::Session;
-use scylla::frame::response::result::CqlValue;
-use scylla::value::CqlDuration;
+use scylla::value::{CqlDuration, CqlValue};
 
 use crate::utils::{create_new_session_builder, setup_tracing, unique_keyspace_name, PerformDDL};
